Split out DataSerializablePrimitive interface.
